+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.22.0 (xxxx-xx-xx)
 
+- Improvement: repeat voteban against the same message cannot be started within 2 minutes
+
 ## 1.21.0 (2024-03-28)
 
 - Improvement: the profanity filter has become more accurate